Struct sixtyfps_interpreter::ComponentDefinition

pub struct ComponentDefinition { /* fields omitted */ }

ComponentDefinition is a representation of a compiled component from .60 markup.

It can be constructed from a .60 file using the ComponentCompiler::build_from_path or ComponentCompiler::build_from_source functions. And then it can be instantiated with the Self::create function.

The ComponentDefinition acts as a factory to create new instances. When you’ve finished creating the instances it is safe to drop the ComponentDefinition.

Implementations

impl ComponentDefinition

pub fn create(&self) -> ComponentInstance

Creates a new instance of the component and returns a shared handle to it.

pub fn properties(&self) -> impl Iterator<Item = (String, ValueType)> + '_

List of publicly declared properties.

pub fn callbacks(&self) -> impl Iterator<Item = String> + '_

Returns the names of all publicly declared callbacks.

pub fn globals(&self) -> impl Iterator<Item = String> + '_

Returns the names of all exported global singletons

pub fn global_properties(
&self,
global_name: &str
) -> Option<impl Iterator<Item = (String, ValueType)> + '_>

List of publicly declared properties in the exported global singleton specified by its name.

pub fn global_callbacks(
&self,
global_name: &str
) -> Option<impl Iterator<Item = String> + '_>

List of publicly declared callbacks in the exported global singleton specified by its name.

pub fn name(&self) -> &str

The name of this Component as written in the .60 file