Struct sixtyfps_interpreter::Struct

pub struct Struct(_);

This type represents a runtime instance of structure in .60.

This can either be an instance of a name structure introduced with the struct keyword in the .60 file, or an anonymous struct written with the { key: value, } notation.

It can be constructed with the FromIterator trait, and converted into or from a Value with the From and TryInto trait

use core::convert::TryInto;
// Construct a value from a key/value iterator
let value : Value = [("foo".into(), 45u32.into()), ("bar".into(), true.into())]
    .iter().cloned().collect::<Struct>().into();

// get the properties of a `{ foo: 45, bar: true }`
let s : Struct = value.try_into().unwrap();
assert_eq!(s.get_field("foo").cloned().unwrap().try_into(), Ok(45u32));

Implementations

impl Struct

pub fn get_field(&self, name: &str) -> Option<&Value>

Get the value for a given struct field

pub fn set_field(&mut self, name: String, value: Value)

Set the value of a given struct field

pub fn iter(&self) -> impl Iterator<Item = (&str, &Value)>

Iterate over all the fields in this struct