Enum sixtyfps_interpreter::Value

#[non_exhaustive]
#[repr(C)]
pub enum Value {
    Void,
    Number(f64),
    String(SharedString),
    Bool(bool),
    Array(SharedVector<Value>),
    Model(Rc<dyn Model<Data = Value>>),
    Struct(Struct),
    Brush(Brush),
    // some variants omitted
}

This is a dynamically typed value used in the SixtyFPS interpreter. It can hold a value of different types, and you should use the From or TryInto traits to access the value.

use core::convert::TryInto;
// create a value containing an integer
let v = Value::from(100u32);
assert_eq!(v.try_into(), Ok(100u32));

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Void

There is nothing in this value. That’s the default. For example, a function that do not return a result would return a Value::Void

Number(f64)

An int or a float (this is also used for unit based type such as length or angle)

String(SharedString)

Correspond to the string type in .60

Bool(bool)

Correspond to the bool type in .60

Array(SharedVector<Value>)

An Array in the .60 language.

Model(Rc<dyn Model<Data = Value>>)

A more complex model which is not created by the interpreter itself (Value::Array can also be used for models)

Struct(Struct)

An object

Brush(Brush)

Correspond to brush or color type in .60. For color, this is then a Brush::SolidColor

Implementations

impl Value

pub fn value_type(&self) -> ValueType

Returns the type variant that this value holds without the containing value.